Reorder columns
Status is now much closer to the beginning.
It's not currently the first column, because tree view forces space for
an icon into the first column, and it looks better with the empty space
in the name field.
On the other hand, if I added icons for the different pending states...